# packages/backend/openapi.yaml
openapi: 3.0.3
info:
  title: ID PASS DataCollect Backend API
  description: |
    REST API for the ID PASS DataCollect backend server. The server is the central
    sync hub for offline-first DataCollect clients (mobile, web, admin UI) and
    coordinates external integrations via the adapter registry.

    ## Authentication

    Most endpoints require a JWT access token in the `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    Access tokens are issued by `POST /api/users/login` and expire after one hour.
    Use `POST /api/users/refresh` to get a new access token (and rotated refresh token)
    without re-prompting for credentials — field agents use this to stay signed in
    while offline. Self-service endpoints (`/api/auth/**`) issue separate scoped
    tokens via OTP, national ID, or OIDC exchange.

    ## Multi-tenant (multi-program) support

    All tenant-scoped endpoints accept a `configId` query parameter (or field in
    the request body). Each configuration has its own entity forms, entity data,
    auth providers, and external-sync settings. Non-admin users can only access
    configurations listed in their `tenantIds` claim — the `validateTenantAccess`
    middleware enforces this on every tenant-scoped route.

    ## Synchronization

    The sync endpoints support bidirectional synchronization between clients and
    server, as well as external system integration. Internal sync uses cursor-based
...
    persisted to `sync_events`.
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin and user JWT issued by `/api/users/login` or `/api/users/refresh`
    SelfServiceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Self-service JWT issued by `/api/auth/otp/verify`, `/api/auth/id/verify`, or `/api/auth/oidc/exchange`
  parameters:
    ConfigIdQuery:
      name: configId
      in: query
      required: false
      schema:
        type: string
        default: default
      description: Program (tenant) identifier. Defaults to `default`.
  schemas:
...
    Event:
      type: object
      required: [guid, entityGuid, type, data, timestamp, userId, syncLevel]
      properties:
        guid:
          type: string
          format: uuid
        entityGuid:
          type: string
          description: Target entity GUID (UUID or human-readable ID such as `hh-001`)
        type:
          type: string
          description: Event type (e.g. `create-group`, `update-individual`, `add-member`)
        data:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        userId:
          type: string
        syncLevel:
          type: integer
          enum: [0, 1]
          description: 0 = LOCAL, 1 = SYNCED
        schemaVersion:
          type: integer
...
    EntitySummary:
      type: object
      properties:
        guid:
          type: string
        id:
          type: string
        name:
          type: string
        entityName:
          type: string
        type:
          type: string
          enum: [group, individual, record]
        data:
          type: object
          additionalProperties: true
        memberIds:
          type: array
          items:
            type: string
          description: Present only when `type = group`
        lastUpdated:
          type: string
          format: date-time
...
paths:
...
  /api/entities:
    get:
      tags: [Entities]
      summary: List entities
      description: Returns entity summaries sorted by `lastUpdated` descending. `limit` is capped at 1000.
      parameters:
        - $ref: '#/components/parameters/ConfigIdQuery'
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 1000 }
      responses:
        '200':
          description: Entity list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EntitySummary' }
        '404':
          description: Tenant not found
  /api/entities/count:
    get:
      tags: [Entities]
      summary: Count entities
      parameters:
        - $ref: '#/components/parameters/ConfigIdQuery'
      responses:
        '200':
          description: Count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
        '404':
          description: Tenant not found
  /api/entities/count-by-form:
    get:
      tags: [Entities]
      summary: Count entities grouped by form
      description: Groups entities by `data.entityName`. Entities without an `entityName` are bucketed under `Unknown`.
      parameters:
        - $ref: '#/components/parameters/ConfigIdQuery'
      responses:
        '200':
          description: Map of form name to count
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { type: integer }
  /api/entities/search:
    post:
      tags: [Entities]
      summary: Search entities
      description: |
        Accepts an array of criteria (max 10) where each criterion is an object whose fields are
        AND'd. String values are capped at 200 characters. `limit` is capped at 500.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                configId: { type: string, default: default }
                criteria:
                  type: array
                  maxItems: 10
                  items: { type: object, additionalProperties: true }
                limit:
                  type: integer
                  default: 50
                  maximum: 500
      responses:
        '200':
          description: Matching entities
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EntitySummary' }
        '400':
          description: Invalid criteria
        '404':
          description: Tenant not found
  /api/entities/{guid}:
    get:
      tags: [Entities]
      summary: Get entity by GUID
      parameters:
        - name: guid
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/ConfigIdQuery'
      responses:
        '200':
          description: Entity
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntitySummary' }
        '404':
          description: Entity or tenant not found
  /api/entities/{guid}/members:
    get:
      tags: [Entities]
      summary: List members of a group
      description: Returns an empty array if the GUID is not a group or has no members. 404 if the GUID does not exist.
      parameters:
        - name: guid
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/ConfigIdQuery'
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EntitySummary' }
        '404':
          description: Entity not found or not a group
  /api/entities/{guid}/events:
    get:
      tags: [Entities]
      summary: List events for an entity
      description: Returns all events for the entity, sorted by timestamp ascending.
      parameters:
        - name: guid
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/ConfigIdQuery'
      responses:
        '200':
          description: Events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Event' }
        '404':
          description: Tenant not found
...
security:
  - BearerAuth: []